@@ -117,7 +117,7 @@ class CreateDraftHandler
     public function markFieldsAsChanged(ElementInterface $element): void
     {
         $fieldLayout = $element->getFieldLayout();
-        $fields = $fieldLayout ? $fieldLayout->getFields() : [];
+        $fields = $fieldLayout ? $fieldLayout->getCustomFields() : [];
 
         foreach ($fields as $field) {
             if (
@@ -161,6 +161,7 @@ class CreateDraftHandler
             'elementId' => $element->getId(),
             'siteId' => $element->getSite()->id,
             'fieldId' => $field->id,
+            'propagated' => $element->propagating,
         ];
 
         $update = [
@@ -188,6 +189,7 @@ class CreateDraftHandler
                 'elementId' => $element->id,
                 'siteId' => $element->siteId,
                 'attribute' => $attribute,
+                'propagated' => $element->propagating,
             ];
 
             $update = [
